+++ framework/helpers/BaseHtml.php
@@ -1093,6 +1093,7 @@ class BaseHtml
      * about attribute expression.
      * @param array $options the tag options in terms of name-value pairs. These will be rendered as
      * the attributes of the resulting tag. The values will be HTML-encoded using [[encode()]].
+     * @param bool $withoutFor the flag set
      * If a value is null, the corresponding attribute will not be rendered.
      * The following options are specially handled:
      *
@@ -1109,6 +1110,7 @@ class BaseHtml
         $for = ArrayHelper::remove($options, 'for', static::getInputId($model, $attribute));
         $attribute = static::getAttributeName($attribute);
         $label = ArrayHelper::remove($options, 'label', static::encode($model->getAttributeLabel($attribute)));
+
         return static::label($label, $for, $options);
     }